+++ src/briefy/leica/tasks/__init__.py
@@ -7,6 +7,7 @@ from briefy.leica.db import Session
 from briefy.leica.log import tasks_logger as logger
 from briefy.leica.tasks.assignment import move_assignments_awaiting_assets
 from briefy.leica.tasks.assignment import notify_late_submissions
+from briefy.leica.tasks.assignment import notify_24hs_shooting
 from briefy.leica.tasks.order import move_orders_accepted
 from briefy.leica.tasks.pool import move_assignments_to_pool
 
@@ -32,9 +33,14 @@ def main():
             logger.info('End: moving assignments to Awaiting Assets.')
 
         with transaction.manager:
-            logger.info('Start: notifying Assignment late submissions.')
+            logger.info('Start: notifying assignments 24hs before shooting.')
+            notify_24hs_shooting()
+            logger.info('End: notifying assignments 24hs before shooting.')
+
+        with transaction.manager:
+            logger.info('Start: notifying assignments not submitted 48hs after shooting.')
             notify_late_submissions()
-            logger.info('End: notifying Assignment late submissions..')
+            logger.info('End: notifying assignments not submitted 48hs after shooting.')
 
         with transaction.manager:
             logger.info('Start: moving orders to accepted.')