+++ tests/tasks/test_task_assignment_notify_before_shooting.py
@@ -1,16 +1,17 @@
-"""Test Task to notify assignment late submission."""
+"""Test Task to notify assignments 24hs before shooting."""
 from briefy.leica import models
-from briefy.leica.tasks.assignment import _notify_late_submissions
-from briefy.leica.tasks.assignment import LATE_SUBMISSION_MSG
-from briefy.leica.tasks.assignment import notify_late_submissions
+from briefy.leica.tasks.assignment import BEFORE_SHOOTING_SECONDS
+from briefy.leica.tasks.assignment import _notify_24hs_shooting
+from briefy.leica.tasks.assignment import BEFORE_SHOOTING_MSG
+from briefy.leica.tasks.assignment import notify_24hs_shooting
 from conftest import BaseTaskTest
 from datetime import timedelta
 
 import json
 
 
-class TestMoveAssignmentToAwaitingAssets(BaseTaskTest):
-    """Test Move Assignment to Awaiting Assets task."""
+class TestNotifyAssignmentBeforeShooting(BaseTaskTest):
+    """Test notify Assignment 24hs before shooting."""
 
     dependencies = [
         (models.Professional, 'data/professionals.json'),
@@ -24,14 +25,23 @@ class TestMoveAssignmentToAwaitingAssets(BaseTaskTest):
     file_path = 'data/assignments.json'
     model = models.Assignment
 
-    def test_wrong_state(self, instance_obj, now_utc):
+    def test_wrong_state_scheduled_datetime(self, instance_obj, now_utc):
         """Will ignore assignment because its state and scheduled_datetime is not the expected one.
 
-        Assignment State is not awaiting assets.
+        Assignment State is not scheduled.
         """
         assignment = instance_obj
-        assignment.scheduled_datetime = now_utc
-        notify_late_submissions()
+        assignment.scheduled_datetime = now_utc + timedelta(seconds=3600)
+        notify_24hs_shooting()
+
+        messages = self.get_messages_from_queue()
+        assert len(messages) == 0
+
+        assignment.scheduled_datetime = now_utc - timedelta(
+            seconds=int(BEFORE_SHOOTING_SECONDS) + 3600
+        )
+        assignment.state = 'pending'
+        notify_24hs_shooting()
 
         messages = self.get_messages_from_queue()
         assert len(messages) == 0
@@ -46,15 +56,17 @@ class TestMoveAssignmentToAwaitingAssets(BaseTaskTest):
         assignment_id = assignment.id
 
         # do not notify with wrong status or scheduled datetime
-        assignment.scheduled_datetime = now_utc
-        assignment.state = 'scheduled'
-        status = _notify_late_submissions(assignment=assignment)
+        assignment.scheduled_datetime = now_utc + timedelta(seconds=3600)
+        assignment.state = 'pending'
+        status = _notify_24hs_shooting(assignment)
         assert status is False
 
         # now we make sure status and scheduled datetime is correct
-        assignment.state = 'awaiting_assets'
-        assignment.scheduled_datetime = now_utc - timedelta(seconds=49 * 3600)
-        status = _notify_late_submissions(assignment=assignment)
+        assignment.state = 'scheduled'
+        assignment.scheduled_datetime = now_utc - timedelta(
+            seconds=int(BEFORE_SHOOTING_SECONDS) - 3600
+        )
+        status = _notify_24hs_shooting(assignment)
 
         assert status is True
         messages = self.get_messages_from_queue()
@@ -62,12 +74,12 @@ class TestMoveAssignmentToAwaitingAssets(BaseTaskTest):
         for message in messages:
             body = json.loads(message.body)
             assert body['event_name'] in (
-                'leica.task.notify_late_submission.success'
+                'leica.task.notify_24hs_before_shooting.success'
             )
 
         assignment = models.Assignment.get(assignment_id)
-        assert assignment.comments[0].content == LATE_SUBMISSION_MSG
+        assert assignment.comments[0].content == BEFORE_SHOOTING_MSG
 
         # assert next time it will not notify again
-        status = _notify_late_submissions(assignment=assignment)
+        status = _notify_24hs_shooting(assignment)
         assert status is False